created the card dealing methods, altering the methods slightly from which was imported. Cards are displayed and the equation is calculated by calculate method then returned to variable to compare against the users input. This is done for both dealing 3 and 5 cards

+++ MathsTutor/RangeValidation.cs
@@ -29,7 +29,7 @@ namespace MathsTutor
             return isValid;
         }
 
-        public bool ValidateRange(float value, float lowerBoundary, float upperBoundary)
+        public bool ValidateRange(double value, double lowerBoundary, double upperBoundary)
         {
             bool isValid = false;
             if (value >= lowerBoundary && value <= upperBoundary)
